Added onclick navigation

+++ app/src/main/java/com/jumptuck/recipebrowser2/singlerecipe/SingleRecipeViewModel.kt
@@ -3,7 +3,7 @@ package com.jumptuck.recipebrowser2.singlerecipe
 import androidx.lifecycle.ViewModel
 import timber.log.Timber
 
-class SingleRecipeViewModel(recipeID: Int) : ViewModel() {
+class SingleRecipeViewModel(recipeID: Long) : ViewModel() {
     init {
         Timber.i("Recipe Index is $recipeID")
     }
