+++ app/src/main/java/com/jumptuck/recipebrowser2/singlerecipe/SingleRecipeViewModelFactory.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import java.lang.IllegalArgumentException
 
-class SingleRecipeViewModelFactory(private val recipeID: Int) : ViewModelProvider.Factory {
+class SingleRecipeViewModelFactory(private val recipeID: Long) : ViewModelProvider.Factory {
     @Suppress("unchecked_cast")
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(SingleRecipeViewModel::class.java)) {
